requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
oauthlib==3.1.1
pycodestyle==2.8.0
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
toml==0.10.2
tweepy==4.4.0
urllib3==1.26.7
